Time: 14 ms (89.42%), Space: 7.2 MB (78.84%) - LeetHub

+++ 1155-number-of-dice-rolls-with-target-sum/1155-number-of-dice-rolls-with-target-sum.cpp
@@ -24,7 +24,7 @@ public:
         }
 
         // Return the result, cast to integer
-        //return static_cast<int>(dp[n][target]); 
-          return dp[n][target];
+        return static_cast<int>(dp[n][target]); 
+          //return dp[n][target];
     }
 };
